Some prototyping of sbatch
